move test options POCOs around, add validator
