+++ src/OptionsPatternValidation/ServiceCollectionExtensions.cs
@@ -1,11 +1,12 @@
 using System.Reflection;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using OptionsPatternValidation.Extensions;
 
 namespace OptionsPatternValidation
 {
-    public static class IServiceCollectionExtensions
+    public static class ServiceCollectionExtensions
     {
         /// <summary>Bind a section of the appsettings.json to an options pattern POCO without validation.</summary>
         public static IServiceCollection AddSettings<T>(
@@ -38,6 +39,27 @@ namespace OptionsPatternValidation
                 .Bind(configurationSection)
                 .RecursivelyValidateDataAnnotations();
             
+            return services;
+        }
+        
+        /// <summary>Bind a section of the appsettings.json to a POCO along with wiring up
+        /// IValidateOptions<T> validation.</summary>
+        public static IServiceCollection AddValidatedSettings<T, TValidator>(
+            this IServiceCollection services,
+            IConfiguration configuration
+            )
+            where T : class, new()
+            where TValidator : class, IValidateOptions<T>
+        {
+            var sectionName = GetSettingsSectionName<T>();
+
+            var configurationSection = configuration.GetSection(sectionName);
+
+            services.AddOptions<T>()
+                .Bind(configurationSection);
+
+            services.AddSingleton<IValidateOptions<T>, TValidator>();
+
             return services;
         }
