+++ test/OptionsPatternValidation.Tests/TestSettings/SimpleSettings.cs
@@ -1,0 +1,9 @@
+namespace OptionsPatternValidation.Tests.TestSettings
+{
+    [SettingsSectionName("Simple")]
+    public class SimpleSettings
+    {
+        public int? IntegerA { get; set; }
+        public bool? BooleanB { get; set; }        
+    }
+}
